Integration Tests - Refactor.
git-svn-id: http://primefaces.googlecode.com/svn/examples/trunk/showcase@6615 5a70e874-a5cb-11dd-8af2-493fb0bb9f7c

+++ src/test/java/com/prime/showcase/integration/AbstractIntegrationTest.java
@@ -217,7 +217,7 @@ public abstract class AbstractIntegrationTest {
 	}
     
     protected Integer getAnimationQueueSizeBySelector(String selector, String queue) {
-        return (Integer) executeJS("return $('" + selector + "').queue('" + queue + "').length;");
+        return ((Long) executeJS("return $('" + selector + "').queue('" + queue + "').length;")).intValue();
     }
     
     protected Boolean anyAnimationInProgress(String selector, String queue) {
